Move argparse utilities to front.py
